+++ frb_codegen/src/library/codegen/parser/mir/parser/mod.rs
@@ -4,3 +4,4 @@ pub(crate) mod function;
 pub(crate) mod misc;
 pub(crate) mod trait_impl;
 pub(crate) mod ty;
+mod function;
